requirements.txt
-e .
absl-py==0.9.0
astor==0.8.1; python_version >= '2.7' and python_version not in '3.0, 3.1, 3.2, 3.3'
astroid==2.4.1
astunparse==1.6.3
attrs==19.3.0
backcall==0.2.0
bleach==3.1.5
blinker==1.4
brotlipy==0.7.0
certifi==2020.6.20
chardet==3.0.4
click==7.1.2
cycler==0.10.0
decorator==4.4.2
defusedxml==0.6.0
entrypoints==0.3
event-service-utils==1.2.6
gast==0.3.3; python_version >= '2.7' and python_version not in '3.0, 3.1, 3.2, 3.3'
google-auth-oauthlib==0.4.1
google-pasta==0.2.0
h5py==2.10.0
importlib-metadata==1.6.0; python_version < '3.8'
ipython-genutils==0.2.0
ipywidgets==7.5.1
isort==4.3.21
jaeger-client==4.1.0
jedi==0.17.0
Jinja2==2.11.2
jsonschema==3.2.0
jupyter==1.0.0
Keras-Preprocessing==1.1.0
kiwisolver==1.2.0
lazy-object-proxy==1.4.3
logzero==1.5.0
markdown==3.2.2; python_version >= '3.5'
MarkupSafe==1.1.1
matplotlib==3.3.1
mccabe==0.6.1
mistune==0.8.4
nbconvert==5.6.1
nbformat==5.0.7
numpy==1.19.1
oauthlib==3.0.1
olefile==0.46
opentracing==2.1.0
packaging==20.4
pandocfilters==1.4.2
pexpect==4.8.0
pickleshare==0.7.5
pillow==5.4.1
prometheus-client==0.8.0
prompt-toolkit==3.0.5
protobuf==3.13.0
ptyprocess==0.6.0
pyasn1==0.4.8
pyasn1-modules==0.2.7
Pygments==2.6.1
PyJWT==1.7.1
pylint==2.5.2
pyOpenSSL==19.1.0
pyparsing==2.4.7
pyrsistent==0.16.0
PySocks==1.7.1
python-dateutil==2.8.1
python-decouple==3.1
pyzmq==19.0.1
QtPy==1.9.0
redis==3.5.3; python_version >= '2.7' and python_version not in '3.0, 3.1, 3.2, 3.3, 3.4'
Send2Trash==1.5.0
sip==4.19.24
six==1.15.0; python_version >= '2.7' and python_version not in '3.0, 3.1, 3.2, 3.3'
tensorboard==2.3.0
tensorflow==2.2.0
termcolor==1.1.0
terminado==0.8.3
testpath==0.4.4
threadloop==1.0.2
thrift==0.13.0
toml==0.10.1
tornado==6.0.4
traitlets==4.3.3
walrus==0.7.1
webencodings==0.5.1
Werkzeug==1.0.1
wheel==0.34.2; python_version >= '3'
widgetsnbextension==3.5.1
wrapt==1.12.1
zipp==3.1.0; python_version >= '3.6'
